♻️ move addEventListener to @datadog/js-core/dom
Move `addEventListener`, `addEventListeners`, `DOM_EVENT`, `isEventSupported`,
`setAllowUntrustedEvents`, `resetAllowUntrustedEvents`, `TrustableEvent`,
and the `VisualViewport` types from browser-core to js-core.

The `noop` dependency is inlined. The spec is moved to js-core as well.

All consumers (browser-core internals, browser-rum-core, browser-rum,
browser-debugger) now import directly from `@datadog/js-core/dom`.
