+++ packages/browser-core/src/browser/webSocketObservable.spec.ts
@@ -1,6 +1,6 @@
 import type { Subscription } from '@datadog/js-core/util'
+import { setAllowUntrustedEvents } from '@datadog/js-core/dom'
 import { registerCleanupTask } from '../../test'
-import { setAllowUntrustedEvents } from './addEventListener'
 import type { WebSocketContext } from './webSocketObservable'
 import { initWebSocketObservable, resetWebSocketObservable } from './webSocketObservable'
