+++ packages/browser-core/src/browser/xhrObservable.ts
@@ -1,9 +1,9 @@
 import { elapsed, timeStampNow, clocksNow } from '@datadog/js-core/time'
 import type { Duration, ClocksState } from '@datadog/js-core/time'
 import { normalizeUrl, globalObject, Observable } from '@datadog/js-core/util'
+import { addEventListener } from '@datadog/js-core/dom'
 import type { InstrumentedMethodCall } from '../tools/instrumentMethod'
 import { instrumentMethod } from '../tools/instrumentMethod'
-import { addEventListener } from './addEventListener'
 
 export interface XhrOpenContext {
   state: 'open'
