+++ packages/browser-core/src/domain/context/storeContextManager.spec.ts
@@ -1,6 +1,6 @@
 import type { Context } from '@datadog/js-core/util'
+import { DOM_EVENT } from '@datadog/js-core/dom'
 import { createNewEvent } from '../../../test'
-import { DOM_EVENT } from '../../browser/addEventListener'
 import { storeContextManager, buildStorageKey, removeStorageListeners } from './storeContextManager'
 import { CustomerDataType } from './contextConstants'
 import { createContextManager } from './contextManager'
