+++ packages/browser-core/src/domain/error/trackRuntimeError.ts
@@ -1,9 +1,9 @@
 import { clocksNow } from '@datadog/js-core/time'
 import { globalObject, Observable } from '@datadog/js-core/util'
+import { DOM_EVENT, isEventSupported } from '@datadog/js-core/dom'
 import { instrumentMethod } from '../../tools/instrumentMethod'
 import type { StackTrace } from '../../tools/stackTrace/computeStackTrace'
 import { computeStackTraceFromOnErrorMessage } from '../../tools/stackTrace/computeStackTrace'
-import { DOM_EVENT, isEventSupported } from '../../browser/addEventListener'
 import { noop } from '../../tools/utils/functionUtils'
 import { computeRawError, isError } from './error'
 import type { RawError } from './error.types'
