+++ packages/browser-core/src/domain/report/reportObservable.ts
@@ -1,8 +1,8 @@
 import { clocksNow } from '@datadog/js-core/time'
 import { monitor } from '@datadog/js-core/monitor'
 import { mergeObservables, Observable } from '@datadog/js-core/util'
+import { addEventListener, DOM_EVENT, isEventSupported } from '@datadog/js-core/dom'
 import { toStackTraceString } from '../../tools/stackTrace/handlingStack'
-import { addEventListener, DOM_EVENT, isEventSupported } from '../../browser/addEventListener'
 import { safeTruncate } from '../../tools/utils/stringUtils'
 import type { RawError } from '../error/error.types'
 import { ErrorHandling, ErrorSource } from '../error/error.types'
