+++ packages/browser-core/src/domain/session/storeStrategies/sessionInLocalStorage.ts
@@ -1,6 +1,6 @@
 import { Observable } from '@datadog/js-core/util'
+import { addEventListener } from '@datadog/js-core/dom'
 import { generateUUID } from '../../../tools/utils/stringUtils'
-import { addEventListener } from '../../../browser/addEventListener'
 import { SessionPersistence } from '../sessionConstants'
 import type { SessionState } from '../sessionState'
 import { isSessionInNotStartedState, toSessionString, toSessionState } from '../sessionState'
