+++ packages/browser-logs/src/boot/preStartLogs.ts
@@ -19,10 +19,10 @@ import {
   startTelemetry,
   TelemetryService,
   startTelemetrySessionContext,
-  setAllowUntrustedEvents,
   isAllowedTrackingOrigins,
 } from '@datadog/browser-core'
 import { BufferedObservable, mockable } from '@datadog/js-core/util'
+import { setAllowUntrustedEvents } from '@datadog/js-core/dom'
 import type { Hooks } from '../domain/hooks'
 import { createHooks } from '../domain/hooks'
 import type { LogsConfiguration, LogsInitConfiguration } from '../domain/configuration'
