+++ packages/browser-rum-core/src/boot/preStartRum.ts
@@ -1,4 +1,5 @@
 import { isWorkerEnvironment, mockable, BufferedObservable } from '@datadog/js-core/util'
+import { setAllowUntrustedEvents } from '@datadog/js-core/dom'
 import { timeStampNow, clocksNow } from '@datadog/js-core/time'
 import { monitorError } from '@datadog/js-core/monitor'
 import type { TimeStamp } from '@datadog/js-core/time'
@@ -25,7 +26,6 @@ import {
   TelemetryService,
   startTelemetrySessionContext,
   addTelemetryDebug,
-  setAllowUntrustedEvents,
   isAllowedTrackingOrigins,
 } from '@datadog/browser-core'
 import type { Hooks } from '../domain/hooks'
