+++ packages/browser-rum-core/src/browser/cookieObservable.ts
@@ -1,6 +1,7 @@
 import { globalObject, setInterval, clearInterval, Observable } from '@datadog/js-core/util'
+import { addEventListener, DOM_EVENT } from '@datadog/js-core/dom'
 import { ONE_SECOND } from '@datadog/js-core/time'
-import { addEventListener, DOM_EVENT, getCookie, isCookieStoreSupported } from '@datadog/browser-core'
+import { getCookie, isCookieStoreSupported } from '@datadog/browser-core'
 
 export type CookieObservable = ReturnType<typeof createCookieObservable>
