+++ packages/browser-rum-core/src/browser/locationChangeObservable.ts
@@ -1,5 +1,6 @@
 import { globalObject, mockable, Observable } from '@datadog/js-core/util'
-import { addEventListener, DOM_EVENT, instrumentMethod, shallowClone } from '@datadog/browser-core'
+import { addEventListener, DOM_EVENT } from '@datadog/js-core/dom'
+import { instrumentMethod, shallowClone } from '@datadog/browser-core'
 
 export interface LocationChange {
   oldLocation: Readonly<Location>
