+++ packages/browser-rum-core/src/browser/viewportObservable.ts
@@ -1,5 +1,6 @@
-import { throttle, addEventListener, DOM_EVENT } from '@datadog/browser-core'
+import { throttle } from '@datadog/browser-core'
 import { Observable } from '@datadog/js-core/util'
+import { addEventListener, DOM_EVENT } from '@datadog/js-core/dom'
 
 export interface ViewportDimension {
   height: number
