+++ packages/browser-rum-core/src/domain/view/bfCacheSupport.ts
@@ -1,4 +1,4 @@
-import { addEventListener, DOM_EVENT } from '@datadog/browser-core'
+import { addEventListener, DOM_EVENT } from '@datadog/js-core/dom'
 
 export function onBFCacheRestore(callback: (event: PageTransitionEvent) => void): () => void {
   const { stop } = addEventListener(
