+++ packages/browser-rum-core/src/domain/view/viewMetrics/trackFirstHidden.ts
@@ -1,5 +1,6 @@
 import type { ClocksState, RelativeTime } from '@datadog/js-core/time'
-import { addEventListeners, DOM_EVENT, noop } from '@datadog/browser-core'
+import { noop } from '@datadog/browser-core'
+import { addEventListeners, DOM_EVENT } from '@datadog/js-core/dom'
 import { supportPerformanceTimingEvent, RumPerformanceEntryType } from '../../../browser/performanceObservable'
 
 export type FirstHidden = ReturnType<typeof trackFirstHidden>
