+++ packages/browser-rum-core/src/domain/view/viewMetrics/trackLargestContentfulPaint.spec.ts
@@ -1,6 +1,6 @@
 import type { RelativeTime } from '@datadog/js-core/time'
 import { clocksOrigin } from '@datadog/js-core/time'
-import { DOM_EVENT } from '@datadog/browser-core'
+import { DOM_EVENT } from '@datadog/js-core/dom'
 import type { Clock } from '@datadog/browser-core/test'
 import {
   setPageVisibility,
