+++ packages/browser-rum-core/src/domain/view/viewMetrics/trackLargestContentfulPaint.ts
@@ -1,6 +1,7 @@
 import type { RelativeTime } from '@datadog/js-core/time'
 import { ONE_MINUTE } from '@datadog/js-core/time'
-import { DOM_EVENT, addEventListeners, findLast } from '@datadog/browser-core'
+import { findLast } from '@datadog/browser-core'
+import { DOM_EVENT, addEventListeners } from '@datadog/js-core/dom'
 import type { RumConfiguration } from '../../configuration'
 import { createPerformanceObservable, RumPerformanceEntryType } from '../../../browser/performanceObservable'
 import { findLcpResourceEntry, getNavigationEntry, sanitizeFirstByte } from '../../../browser/performanceUtils'
