+++ packages/browser-rum-core/src/domain/view/viewMetrics/trackScrollMetrics.ts
@@ -1,8 +1,9 @@
 import { ONE_SECOND, elapsed, relativeNow } from '@datadog/js-core/time'
 import type { Duration, ClocksState } from '@datadog/js-core/time'
 import { monitor } from '@datadog/js-core/monitor'
-import { throttle, addEventListener, DOM_EVENT } from '@datadog/browser-core'
+import { throttle } from '@datadog/browser-core'
 import { Observable } from '@datadog/js-core/util'
+import { addEventListener, DOM_EVENT } from '@datadog/js-core/dom'
 import { getScrollY } from '../../../browser/scroll'
 import { getViewportDimension } from '../../../browser/viewportObservable'
