+++ packages/browser-rum/src/domain/record/trackers/trackFocus.ts
@@ -1,4 +1,4 @@
-import { DOM_EVENT, addEventListeners } from '@datadog/browser-core'
+import { DOM_EVENT, addEventListeners } from '@datadog/js-core/dom'
 import { timeStampNow } from '@datadog/js-core/time'
 import type { FocusRecord } from '../../../types'
 import { RecordType } from '../../../types'
