+++ packages/browser-rum/src/domain/record/trackers/trackMouseInteraction.spec.ts
@@ -1,4 +1,4 @@
-import { DOM_EVENT } from '@datadog/browser-core'
+import { DOM_EVENT } from '@datadog/js-core/dom'
 import { createNewEvent, registerCleanupTask } from '@datadog/browser-core/test'
 import { appendElement } from '../../../../../browser-rum-core/test'
 import { IncrementalSource, MouseInteractionType, RecordType } from '../../../types'
