+++ packages/browser-rum/src/domain/record/trackers/trackMove.ts
@@ -1,4 +1,5 @@
-import { addEventListeners, DOM_EVENT, throttle } from '@datadog/browser-core'
+import { throttle } from '@datadog/browser-core'
+import { addEventListeners, DOM_EVENT } from '@datadog/js-core/dom'
 import type { BrowserIncrementalSnapshotRecord, MousemoveData, MousePosition } from '../../../types'
 import { IncrementalSource } from '../../../types'
 import { getEventTarget, isTouchEvent } from '../eventsUtils'
