+++ packages/browser-rum/src/domain/record/trackers/trackViewportResize.ts
@@ -1,4 +1,5 @@
-import { throttle, DOM_EVENT, addEventListeners, noop } from '@datadog/browser-core'
+import { throttle, noop } from '@datadog/browser-core'
+import { DOM_EVENT, addEventListeners } from '@datadog/js-core/dom'
 import { timeStampNow } from '@datadog/js-core/time'
 import type { ViewportDimension } from '@datadog/browser-rum-core'
 import { initViewportObservable } from '@datadog/browser-rum-core'
