+++ packages/js-core/src/dom/addEventListener.spec.ts
@@ -1,6 +1,6 @@
-import { createNewEvent, mockZoneJs, registerCleanupTask } from '../../test'
-import type { MockZoneJs } from '../../test'
-import { noop } from '../tools/utils/functionUtils'
+import { createNewEvent, mockZoneJs, registerCleanupTask } from '../../../browser-core/test'
+import type { MockZoneJs } from '../../../browser-core/test'
+import { noop } from '../util/noop'
 import { addEventListener, DOM_EVENT, isEventSupported, setAllowUntrustedEvents } from './addEventListener'
 
 describe('addEventListener', () => {
