+++ packages/js-core/src/entries/dom.ts
@@ -1,1 +1,13 @@
 export { fetch } from '../dom/fetch'
+
+export type { TrustableEvent } from '../dom/addEventListener'
+
+export { DOM_EVENT } from '../dom/addEventListener'
+
+export {
+  addEventListener,
+  addEventListeners,
+  isEventSupported,
+  setAllowUntrustedEvents,
+  resetAllowUntrustedEvents,
+} from '../dom/addEventListener'
