+++ tsconfig.base.json
@@ -57,7 +57,8 @@
       "@datadog/js-core/util": ["./packages/js-core/src/entries/util"],
       "@datadog/js-core/transport": ["./packages/js-core/src/entries/transport"],
       "@datadog/js-core/sample": ["./packages/js-core/src/entries/sample"],
-      "@datadog/js-core/configuration": ["./packages/js-core/src/entries/configuration"]
+      "@datadog/js-core/configuration": ["./packages/js-core/src/entries/configuration"],
+      "@datadog/js-core/dom": ["./packages/js-core/src/entries/dom"]
     }
   }
 }
